feat: added shell of MainPage, CoursePage and QuizPage
Added Pages and their complex components. Not connected to api, for now operating on sample data.

Some functionality is locked to a logged in user. Since it is not possible to properly log in, to see full functionality some code is to be disabled.

MainPage: filters, search bar, course selection with navigation
CoursePage: Teacher card with week's availability and other courses (with navigation based on ids). Course details, variant selection, class setup (with available slot selection for the upcoming month). Teacher opinions.
QuizPage: basic filters and search (not functional), interfaces for components. Very early state.